pyproject.toml
[project]
name = "telegram-downloader"
version = "0.1.0"
description = "Telegram Bot to download video files sent to it"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pydantic-settings>=2.7.0",
    "pydantic>=2.10.3",
    "python-dotenv>=1.0.1",
    "python-telegram-bot>=21.9",
]

[tool.uv]
dev-dependencies = [
    "ruff>=0.8.3",
]
